data_analysis_chatbot/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
plotly>=5.3.0

# NLP and RAG
sentence-transformers>=2.2.0
openai>=1.0.0
langchain>=0.0.200
langchain-openai>=0.0.1
langchain-core>=0.0.10
langchain-community>=0.0.10
pydantic>=2.0.0
faiss-cpu>=1.7.0
nltk>=3.6.0
transformers>=4.18.0

# Database
sqlalchemy>=2.0.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
pymongo>=4.0.0
redis>=4.0.0
chromadb>=0.4.0

# API
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=0.19.0
httpx>=0.24.0

# UI
streamlit>=1.20.0
streamlit-chat>=0.0.2.2
streamlit-extras>=0.2.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
tqdm>=4.62.0
loguru>=0.6.0
python-multipart>=0.0.5
aiofiles>=0.8.0